+++ spec/proxy_spec.rb
@@ -12,6 +12,7 @@ describe 'openstack-object-storage::proxy-server' do
       swift_stubs
       @chef_run = ::ChefSpec::Runner.new ::UBUNTU_OPTS
       @node = @chef_run.node
+      @node.set['cpu']['total'] = 6
       @node.set['lsb']['code'] = 'precise'
       @node.set['swift']['authmode'] = 'swauth'
       @node.set['swift']['platform']['swauth_packages'] = ['swauth']
@@ -52,12 +53,72 @@ describe 'openstack-object-storage::proxy-server' do
         expect(sprintf("%o", @file.mode)).to eq "600"
       end
 
-      it "template contents" do
-        pending "TODO: implement"
+      it "has proper pipeline in template" do
+        array = [
+        /^pipeline = catch_errors healthcheck cache ratelimit swauth proxy-logging proxy-server$/,
+        /^workers = 5$/,
+        ]
+        array.each do |content|
+          expect(@chef_run).to render_file(@file.name).with_content(content)
+        end
       end
 
     end
 
-  end
+    describe "/etc/swift/proxy-server.conf with domain_remap enabled" do
 
+      before do
+        @node = @chef_run.node
+        @node.set["swift"]["domain_remap"]["enabled"] = true
+        @chef_run.converge "openstack-object-storage::proxy-server"
+        @file = @chef_run.template "/etc/swift/proxy-server.conf"
+      end
+
+      it "has proper pipeline in template" do
+        array = [
+          /^pipeline = catch_errors healthcheck cache ratelimit domain_remap swauth proxy-logging proxy-server$/
+       ]
+       array.each do |content|
+          expect(@chef_run).to render_file(@file.name).with_content(content)
+        end
+      end
+    end
+
+    describe "/etc/swift/proxy-server.conf with formpost enabled" do
+
+      before do
+        @node = @chef_run.node
+        @node.set["swift"]["formpost"]["enabled"] = true
+        @chef_run.converge "openstack-object-storage::proxy-server"
+        @file = @chef_run.template "/etc/swift/proxy-server.conf"
+      end
+
+      it "has proper pipeline in template" do
+        array = [
+          /^pipeline = catch_errors healthcheck cache ratelimit formpost swauth proxy-logging proxy-server$/
+       ]
+       array.each do |content|
+          expect(@chef_run).to render_file(@file.name).with_content(content)
+        end
+      end
+    end
+    describe "/etc/swift/proxy-server.conf with staticweb enabled" do
+
+      before do
+        @node = @chef_run.node
+        @node.set["swift"]["staticweb"]["enabled"] = true
+        @chef_run.converge "openstack-object-storage::proxy-server"
+        @file = @chef_run.template "/etc/swift/proxy-server.conf"
+      end
+
+      it "has proper pipeline in template" do
+        array = [
+          /^pipeline = catch_errors healthcheck cache ratelimit swauth staticweb proxy-logging proxy-server$/
+       ]
+       array.each do |content|
+          expect(@chef_run).to render_file(@file.name).with_content(content)
+        end
+      end
+    end
+  end
 end